chat: stop console error after clicking marsey emoji speed modal (#119)
Kindly,

As per $title. This is just a suggestion implementation

Whenever we click any marsey in the speed modal, it generates an error in developer tools console because markdown() is not defined. I suspect not being available is expected because that function is for re-rendering the markdown preview (????? when I look at chat code it is all spaghette to me)

Thanks,
granny

Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/119
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>