docker container startup issues: remove public.users cardview from db init (#132)
Kindly,

I think this required follow-up work after either of these:
* https://fsdfsd.net/rDrama/rDrama/commit/7fa150a27f58bf48d71295b9c5afb57311edc272
* https://fsdfsd.net/rDrama/rDrama/commit/8f4ba57b6fadd5e894d159f2be6a97e9c0ba3b22

but it might also be the other way around. I just did this so I could have a docker container that works on my machine (TM). I hadn't restarted it in a few days

Either way, this is a totally irresponsible PR from the granny that is guaranteed to make people thank granny for crashing rdrama.

Thanks,
granny

Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/132
Co-authored-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>
Co-committed-by: mummified-corroding-granny <mummified-corroding-granny@noreply.fsdfsd.net>