Bring back orgies (watchparties), now controllable by admins, and generally better in all ways (#165)
This PR adds orgies back into rdrama. Long ago, snakes made the original orgy code, and it was super fun. But he had to rush it out, and ended up making it a bit unsustainable, and had a couple questionable coding decisions, which meant that it had to be removed. Hey, the man literally did it in a few hours before the DB trial continued, lmao.

Anyways, I took my own approach to it. I do not use iframes, i just just repurpose code from /chat window. Because I had that freedom, I also moved things around to make the user experience a bit better. I also added a title to give users some context about what's happening. Check it out

![image](/attachments/6719146c-4922-4d75-967d-8d424a09b198)

Most importantly, this is all configurable from the site. Admins with the permission "ORGIES" will see this in their control panel

![image](/attachments/423d6046-a11d-4e84-bd2c-a2a641afd552)

Nigga, idk where to put it, so I made my own category.

If there is no orgy in progress, admins will see this:

![image](/attachments/7c64b9fa-cdf4-4986-a0c4-f2324878062e)

Click the button, and, viola, the orgy begins.

If there is an orgy in progress, the page will look like this:

![image](/attachments/b65be4b3-5db1-43cb-8857-7d3a8ea24ca7)

Click the button, and the orgy stops.

If an orgy is in progress, navigating to /chat will take the user to the orgy seemlessly. But what if they don't want to participate, liek some kind of spoilsport? Just navigate to /old_chat.

That's just about it, it's really that simple. I have lots of ideas for the future, but I'll let that wait til later :).

A few notes about implementation:
 - I moved some functionality out of /templates/chat.html and into /templates/util/macros.html. This is just so I could reference the code directly from my new template, /templates/orgy.html.
 - The orgy is stored as a single row in the new table "orgies". Okay, I know this is a little silly, but you know what they say: "if it's stupid and it works, it's not stupid". (tbf the oceangate ceo also said that)

Co-authored-by: Chuck Sneed <sneed@formerlychucks.net>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/165
Co-authored-by: HeyMoon <heymoon@noreply.fsdfsd.net>
Co-committed-by: HeyMoon <heymoon@noreply.fsdfsd.net>