Adds marseylove (marseyheart so it doesn't interfere with marseylove) (#184)
This was the emoji modifier I wanted to add all along, seems to work well with all of the other modifiers.

Made the modifer :-heart: so that it doesn't interfere with all of the current love emotes that everyone put so much work into.

Check out the screenshot for how it looks locally.

Co-authored-by: Chuck <dude@bussy.com>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/184
Co-authored-by: top <top@noreply.fsdfsd.net>
Co-committed-by: top <top@noreply.fsdfsd.net>