Adds the replacement gold animation (#227)
Any chance you could push this to only devrama so I can test some edge cases?

This seems to be really fast, most of the credit goes to @trihard for making the svg, I just did the lazy drop in bit.

Anyways, I don't really have a slow machine to test this on so I'm not sure how much this improves performance but should cut down on repaints

Co-authored-by: Mike Hawk <top@rdrama.net>
Reviewed-on: https://fsdfsd.net/rDrama/rDrama/pulls/227
Co-authored-by: top <top@noreply.fsdfsd.net>
Co-committed-by: top <top@noreply.fsdfsd.net>